fix(contracts): fixed data type in build system

+++ packages/contracts/src/systems/BuildSystem.sol
@@ -23,9 +23,9 @@ uint256 constant ID = uint256(keccak256("system.Build"));
 contract BuildSystem is System {
   constructor(IWorld _world, address _components) System(_world, _components) {}
 
-//Moresh: Updated to support entity_type
+  //Moresh: Updated to support entity_type
   function execute(bytes memory arguments) public returns (bytes memory) {
-    (int32 x, int32 y, int32 entity_type) = abi.decode(arguments, (int32, int32, int32));
+    (int32 x, int32 y, uint256 entity_type) = abi.decode(arguments, (int32, int32, uint256));
 
     require(x >= -25 && x <= 25, "Invalid X co-ordinate");
     require(y >= -25 && y <= 25, "Invalid Y co-ordinate");
@@ -144,7 +144,11 @@ contract BuildSystem is System {
     );
   }
 
-  function executeTyped(int32 x, int32 y, int32 entity_type) public returns (bytes memory) {
+  function executeTyped(
+    int32 x,
+    int32 y,
+    uint256 entity_type
+  ) public returns (bytes memory) {
     return execute(abi.encode(x, y, entity_type));
   }
 }
